Crate cryptoballot

CryptoBallot is a cryptographically secure online voting system, providing secure anonymous voting with end-to-end verifiability.

It is currenly under active development and is not production ready.

CryptoBallot is fundamentally a transaction processor and validator. When transactions are validated in order, it creates an end-to-end verifiable voting system.

Glossary:

Transaction 1: Election Transaction - Defines an election, created by an election authority.

Transaction 2: Vote Transaction - Posted by a voter to cast a vote in an election.

Transaction 3: Secret Share Transaction - Posted by a trustee to allow votes to be decrypted and viewed.

Transaction 4: Decryption Transaction - Decrypt a vote, allowing it to be tallied.

Election Authority - Creates an Election Transaction and distributes the encryption secret to trustees via Shamir Secret Sharing.

Trustee - Holds a vote-decryption secret share, posts Secret Share Transactions.

Authenticator - Certifies that a voter can vote an election and ballot using blind-signing.

Ballot - A set of contests for an election, usually restricted to a geographic area.

Re-exports

pub extern crate cryptid;

pub extern crate ed25519_dalek;

pub extern crate rsa;

pub extern crate uuid;

pub extern crate x25519_dalek;

Structs

AuthPackage
The Auth Package triplet of election-id, ballot-id, and voter public key

AuthPublicKey
RSA Public Key for blind signing

Authentication
An Authentication is returned by an authenticator, clearing the voter to vote.

Authenticator
An Authenticator is responsible for authenticating a voter as allowed to vote a specific ballot in an election.

DecryptionTransaction
Transaction 9: Decryption

ElectionTransaction
Transaction 1: Election

EncryptedShare

EncryptionKeyTransaction
Transaction 4: PublicKeyConfirmationTransaction

Identifier
Transaction identifier

KeyGenCommitmentTransaction
Transaction 2: KeyGenCommitmentTransaction

KeyGenPublicKeyTransaction
Transaction 4: KeyGenPublicKeyTransaction

KeyGenShareTransaction
Transaction 3: KeyGenShareTransaction

MemStore
A simple store that uses an in-memory BTreeMap

MixConfig

MixTransaction

PartialDecryptionTransaction
Transaction 8: Partial Decryption

Signed
A generic signed transaction

TransactionNotFound

Trustee
A trustee is responsible for safeguarding a secret share (a portion of the secret vote decryption key), distributed by the election authority via Shamir Secret Sharing.

VoteTransaction
Transaction 2: Vote

VotingEndTransaction
Transaction 7: VotingEnd

Enums

Error
Error types

SignedTransaction
A signed transaction

Transaction
An unsigned transaction TODO: Implment From going for specific tx to this emum and vice versa

TransactionType
A transaction type

ValidationError
Transaction Validation errors

Traits

CryptoBallotTransaction
All CryptoBallot transactions implement this trait

Store
A transaction store

Functions

decrypt_vote
Decrypt the vote from the given partial decryptions.

encrypt_vote
Encrypt a vote with the public key provided by the encryption_key transaction (EncryptionKeyTransaction.encryption_key)

encrypted_vote_from_upstream_tx
A convenience function for getting an encrypted-vote from some upstream transaction ID. The upstream transaction should either be a mixnet or a vote transaction.

generate_keypair
Generate an ed25519 keypair

mix
Do a mixnet shuffle This is an expensive and time-consuming operation, so should ideally be offloaded to it’s own thread

verify_mix
Verify mixnet shuffle